refac: move and rename the network files to the specified folder
According to the different network types, move and rename the
network files to the specified folder.
